SNOW-1055059: Moving SPCS-specific integration testing utils into the 'spcs' folder.
